LayoutTests/css3/font-feature-settings-parsing-expected.txt

Test parsing of the CSS3 font-feature-settings property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


- Tests valid inputs.
PASS canonicalize(parseResultOf("valid_normal")) is "normal"
PASS canonicalize(parseResultOf("valid_value_1")) is "\"dlig\""
PASS canonicalize(parseResultOf("valid_value_2")) is "\"swsh\" 2"
PASS canonicalize(parseResultOf("valid_value_on")) is "\"smcp\""
PASS canonicalize(parseResultOf("valid_value_off")) is "\"liga\" 0"
PASS canonicalize(parseResultOf("valid_value_omit")) is "\"c2sc\""
PASS canonicalize(parseResultOf("valid_valuelist")) is "\"hist\", \"tnum\""
PASS canonicalize(parseResultOf("valid_singlequote")) is "\"pkrn\""
PASS canonicalize(parseResultOf("valid_unusual_tag")) is "\"!@#$\""
PASS canonicalize(parseResultOf("valid_tag_space")) is "\"a bc\""
PASS canonicalize(parseResultOf("valid_composite")) is "\"dlig\", \"lig \" 0, \"smcp\""
- Tests invalid inputs.  Results should be "normal".
PASS parseResultOf("invalid_ident") is "normal"
PASS parseResultOf("invalid_cases") is "normal"
PASS parseResultOf("invalid_1") is "normal"
PASS parseResultOf("invalid_off") is "normal"
PASS parseResultOf("invalid_normal_duplicate") is "normal"
PASS parseResultOf("invalid_normal_list") is "normal"
PASS parseResultOf("invalid_longer_ident") is "normal"
PASS parseResultOf("invalid_longer_quote") is "normal"
PASS parseResultOf("invalid_negative") is "normal"
PASS parseResultOf("invalid_float") is "normal"
PASS parseResultOf("invalid_missing_comma") is "normal"
PASS parseResultOf("invalid_missing_comma_ident") is "normal"
PASS parseResultOf("invalid_comma") is "normal"
PASS parseResultOf("invalid_ending_comma") is "normal"
PASS parseResultOf("invalid_beginning_comma") is "normal"
PASS parseResultOf("invalid_on") is "normal"
PASS parseResultOf("invalid_0") is "normal"
- Tests inherit.
PASS canonicalize(parseResultOf("outer")) is "\"dlig\""
PASS canonicalize(parseResultOf("inner")) is "\"dlig\""
- Tests @font-face.
PASS canonicalize(fontFaceRuleValid) is "\"liga\""
PASS fontFaceRuleInvalid is ""
PASS successfullyParsed is true

TEST COMPLETE
